+++ desktop/android/app/src/androidTest/java/com/echodesk/app/ExampleInstrumentedTest.java
@@ -21,6 +21,6 @@ public class ExampleInstrumentedTest {
         // Context of the app under test.
         Context appContext = InstrumentationRegistry.getInstrumentation().getTargetContext();
 
-        assertEquals("com.echodesk.app", appContext.getPackageName());
+        assertTrue(appContext.getPackageName().startsWith("com.echodesk."));
     }
 }